Desenvolvimento

+++ Parking.Management.Dashboard/Backend/src/domain/use-cases/course/delete/delete-course-usecase.ts
@@ -8,13 +8,21 @@ export class DeleteCourseUseCase {
   constructor(private courseRepository: ICourseRepository) {}
 
   async execute(id: string): Promise<Result<CourseDto>> {
+    const course = await this.courseRepository.findById(id);
+
+    if (!course) {
+      return new Result<CourseDto>({
+        message: "Curso não encontrada.",
+      });
+    }
+
     const result = await this.courseRepository.delete(id);
 
     const courseDto = mapper.map(result, Course, CourseDto);
 
-    return new Result<CourseDto>(
-      courseDto,
-      "Universidade deletada com sucesso."
-    );
+    return new Result<CourseDto>({
+      content: courseDto,
+      message: "Curso deletado com sucesso.",
+    });
   }
 }
